feat: build BannerCarousel component and homepage banner in progress
