+++ frontend/src/components/ProductCard.vue
@@ -26,5 +26,7 @@ defineProps({
   product: Object
 })
 
-
+const addToCart = (product) => {
+  console.log('加入購物車', product)
+}
 </script>
