+++ frontend/src/main.js
@@ -17,3 +17,4 @@ app.use(PrimeVue, {
 app.use(createPinia())
 app.use(router)
 app.mount('#app')
+
