+++ frontend/src/router/index.js
@@ -7,7 +7,7 @@ const router = createRouter({
   routes: [
     {
       path: '/',
-      name: 'home',
+      name: 'HomeView.vue',
       component: HomeView,
     },
     {
